MDRange: ConstArray and MDRangePolicy function added
Allows user to pass (outer) layout as arg to the MDRangePolicy fcn, and
(inner) layout to the Tile function.
Tested locally on laptop with gcc and clang.
Extra unit tests of the new API added to test_for2 - will not add for
all ranks, as this will increase test time.